refactor: relocate disposable
