+++ src/plugin/devices.ts
@@ -1,5 +1,5 @@
 import { type DeviceInfo } from "../api/device";
-import type { IDisposable } from "./common/disposable";
+import type { IDisposable } from "../common/disposable";
 import type { StreamDeckConnection } from "./connectivity/connection";
 import { DeviceDidConnectEvent, DeviceDidDisconnectEvent, DeviceEvent } from "./events";
